im stupid
"| 1" was written in wrong place (im not that stupid)
but I do forgot to convert the whole thing to int LOL

+++ source/gaussianblur.cpp
@@ -45,7 +45,7 @@ const VSFrame *VS_CC gblurGetFrame(int n, int activationReason, void *instanceDa
             }
 
             if (w != width || h != height) {
-                int sizeXS = (d->sizeX * ((double)w / width) | 1), sizeYS = (d->sizeY * ((double)h / height) | 1);
+                int sizeXS = int(d->sizeX * ((double)w / width)) | 1, sizeYS = int(d->sizeY * ((double)h / height)) | 1;
                 cv::GaussianBlur(inM, outM, cv::Size(sizeXS, sizeYS), d->sigmaX * ((double)w / width), d->sigmaY * ((double)h / height), d->borderType, cv::ALGO_HINT_ACCURATE);
             }
             else
